requirements.txt
ale_py==0.7.4
control==0.9.2
dmc2gym==1.0.0
gym==0.23.1
imageio==2.25.0
matplotlib==3.5.2
moviepy==1.0.3
numpy==1.20.3
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.1.0
Pillow==10.1.0
pybullet==3.2.5
pybulletgym.egg==info
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.0
termcolor==2.3.0
torch==1.11.0
torchvision==0.12.0
